Extract ProcessorContext as top-level class
