+++ presto-main/src/main/java/io/prestosql/operator/WorkProcessorOperatorAdapter.java
@@ -14,8 +14,6 @@
 package io.prestosql.operator;
 
 import com.google.common.util.concurrent.ListenableFuture;
-import io.prestosql.Session;
-import io.prestosql.execution.Lifespan;
 import io.prestosql.memory.context.MemoryTrackingContext;
 import io.prestosql.spi.Page;
 
@@ -117,49 +115,5 @@ public class WorkProcessorOperatorAdapter
             throws Exception
     {
         workProcessorOperator.close();
-    }
-
-    public static class ProcessorContext
-    {
-        private final Session session;
-        private final MemoryTrackingContext memoryTrackingContext;
-        private final DriverYieldSignal driverYieldSignal;
-        private final Lifespan lifespan;
-        private final SpillContext spillContext;
-
-        public ProcessorContext(Session session, MemoryTrackingContext memoryTrackingContext, OperatorContext operatorContext)
-        {
-            this.session = requireNonNull(session, "session is null");
-            this.memoryTrackingContext = requireNonNull(memoryTrackingContext, "memoryTrackingContext is null");
-            requireNonNull(operatorContext, "operatorContext is null");
-            this.driverYieldSignal = operatorContext.getDriverContext().getYieldSignal();
-            this.lifespan = operatorContext.getDriverContext().getLifespan();
-            this.spillContext = operatorContext.getSpillContext();
-        }
-
-        public Session getSession()
-        {
-            return session;
-        }
-
-        public MemoryTrackingContext getMemoryTrackingContext()
-        {
-            return memoryTrackingContext;
-        }
-
-        public DriverYieldSignal getDriverYieldSignal()
-        {
-            return driverYieldSignal;
-        }
-
-        public Lifespan getLifespan()
-        {
-            return lifespan;
-        }
-
-        public SpillContext getSpillContext()
-        {
-            return spillContext;
-        }
     }
 }
